+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Renamed ArcPanel to ChapterArcPanel
 - Renamed BrowseTextBox to ChapterBrowseTextBox
 - Renamed DynamicTabControl to ChapterTabControl
+- Renamed EllipsePanel to ChapterEllipsePanel
 ### Supported .Net Versions
 - .Net Core 3.0
 - .Net Framework 4.5